initial commit for statestore
