+++ store/store_test.go
@@ -8,7 +8,6 @@ import (
 	abcitypes "github.com/tendermint/tendermint/abci/types"
 	tmstate "github.com/tendermint/tendermint/proto/tendermint/state"
 
-	"github.com/celestiaorg/optimint/state"
 	"github.com/celestiaorg/optimint/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -112,71 +111,6 @@ func TestStoreLoad(t *testing.T) {
 			})
 		}
 	}
-}
-
-func TestRestart(t *testing.T) {
-	t.Parallel()
-
-	assert := assert.New(t)
-
-	kv := NewDefaultInMemoryKVStore()
-	s1 := New(kv)
-	expectedHeight := uint64(10)
-	//block := getRandomBlock(expectedHeight, 10)
-	//err := s1.SaveBlock(block, &types.Commit{Height: block.Header.Height, HeaderHash: block.Header.Hash()})
-	err := s1.UpdateState(state.State{
-		LastBlockHeight: int64(expectedHeight),
-	})
-	assert.NoError(err)
-
-	s2 := New(kv)
-	_, err = s2.LoadState()
-	assert.NoError(err)
-
-	assert.Equal(expectedHeight, s2.Height())
-}
-
-func TestBlockResponses(t *testing.T) {
-	t.Parallel()
-	assert := assert.New(t)
-
-	kv := NewDefaultInMemoryKVStore()
-	s := New(kv)
-
-	expected := &tmstate.ABCIResponses{
-		BeginBlock: &abcitypes.ResponseBeginBlock{
-			Events: []abcitypes.Event{{
-				Type: "test",
-				Attributes: []abcitypes.EventAttribute{{
-					Key:   []byte("foo"),
-					Value: []byte("bar"),
-					Index: false,
-				}},
-			}},
-		},
-		DeliverTxs: nil,
-		EndBlock: &abcitypes.ResponseEndBlock{
-			ValidatorUpdates: nil,
-			ConsensusParamUpdates: &abcitypes.ConsensusParams{
-				Block: &abcitypes.BlockParams{
-					MaxBytes: 12345,
-					MaxGas:   678909876,
-				},
-			},
-		},
-	}
-
-	err := s.SaveBlockResponses(1, expected)
-	assert.NoError(err)
-
-	resp, err := s.LoadBlockResponses(123)
-	assert.Error(err)
-	assert.Nil(resp)
-
-	resp, err = s.LoadBlockResponses(1)
-	assert.NoError(err)
-	assert.NotNil(resp)
-	assert.Equal(expected, resp)
 }
 
 func getRandomBlock(height uint64, nTxs int) *types.Block {
